pyproject.toml
[project]
name = "phenocai"
version = "0.1.0"
description = "Automated phenological camera image analysis and classification using AI"
readme = "README.md"
authors = [
    { name = "Jobelund", email = "124563223+jobelund@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.0",
    "duckdb>=1.3.0",
    "marimo>=0.13.11",
    "matplotlib>=3.10.3",
    "numpy>=2.1.3",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "polars>=1.30.0",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "tensorflow>=2.19.0",
]

[project.scripts]
phenocai = "phenocai.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
